+++ src/main/java/chess/web/BoardController.java
@@ -1,18 +1,19 @@
 package chess.web;
 
-import chess.service.GameService;
-import chess.web.dto.BoardDto;
-import chess.web.dto.CommendDto;
-import chess.web.dto.ResultDto;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+import chess.service.GameService;
+import chess.web.dto.BoardDto;
+import chess.web.dto.CommendDto;
+import chess.web.dto.ResultDto;
+
+@RestController
 @RequestMapping("/boards")
 public class BoardController {
