--- src/main/java/chess/web/HomeController.java
@@ -1,13 +1,0 @@
-package chess.web;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String login() {
-        return "login.html";
-    }
-}
