+++ src/main/resources/static/login.html
@@ -80,7 +80,6 @@
                 body: 'password=' + password
             })
                 .then(res =>{
-                    alert("방이 삭제되었습니다.");
                     let roomContainer = document.getElementById('roomNames');
                     roomContainer.innerHTML = "";
                 })
